requirements.txt
scikit-learn
pandas
pillow
numpy
timm
matplotlib
scipy
wfdb
neurokit2
torch 
torchvision 
torchaudio
transformers